create custom modeling mixin
